+++ src/simmate/database/workflow_results/__init__.py
@@ -14,6 +14,7 @@ from simmate.database.base_data_types import (
     StaticEnergy,
 )
 from simmate.toolkit.structure_prediction.evolution.database import (
+    ChemicalSystemSearch,
     FixedCompositionSearch,
     SteadystateSource,
 )